Refactor ChromeWriter class: separate implementation from header and improve event handling
